updated shapes classes

+++ java_II/src/Shapes/Measurable.java
@@ -1,6 +1,6 @@
 package Shapes;
 
 public interface Measurable {
-    double getArea();
-    double getPerimeter();
+    long getArea();
+    long getPerimeter();
 }
